Chore: Refactor secrets service
